+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.17.1 - 2021-10-12
+
++ [[#159]] Add `ValueType::column_type`
++ [[#160]] Add `FromValueTuple` trait
+
+[#159]: https://github.com/SeaQL/sea-query/pull/159
+[#160]: https://github.com/SeaQL/sea-query/pull/160
+
 ## 0.17.0 - 2021-10-06
 
 + [[#157]] Fix binding nullable custom types on db drivers